split up AddingFragmentTest into smaller classes
